Refactor: refactor `ModelStub` out
Refactors `ModelStub` out from `AddCustomerCommandTest` so that other add commands
can also implement the `ModelStub` for testing.
